linked items moved to owner "Recycle Bin" on delete

+++ main/functions.py
@@ -6,7 +6,7 @@ from django.forms import modelform_factory
 from django.shortcuts import redirect
 
 from main.models import Variable, CellModel, Component, Reset, CompoundUnit, Unit, \
-    Math, Prefix
+    Math, Prefix, Person
 
 
 # ---------------------- CELLML FUNCTIONS -------------------------------------
@@ -595,7 +595,7 @@ def get_relationship_menu(item_model):
 
 # ------------------------------- COPY FUNCTIONS ------------------------------------------
 
-def create_by_shallow_copy(request, item, exclude=[]):
+def create_by_shallow_copy(request, item, exclude=[], options=[]):
     """
     This just copies the local attributes of the item from one to another. It is used by both the deep and shallow copy
     functions later on, but the placeholder object creation happens here.
@@ -632,7 +632,7 @@ def create_by_shallow_copy(request, item, exclude=[]):
     return old_item, item
 
 
-def link_copy(request, from_item, to_item, exclude=[]):
+def link_copy(request, from_item, to_item, exclude=[], options=[]):
     """
     Copies the related fields by transferring the existing reference not by creating new objects
     :param request: request (for passing back messages)
@@ -683,7 +683,7 @@ def link_copy(request, from_item, to_item, exclude=[]):
     return from_item, to_item
 
 
-def deep_copy(request, from_item, to_item, exclude=[]):
+def deep_copy(request, from_item, to_item, exclude=[], options=[]):
     """
     Makes a duplicate of all connected items instead of copying them as a link
     :param request:
@@ -707,16 +707,16 @@ def deep_copy(request, from_item, to_item, exclude=[]):
     for f in m2o_fields:
         for related_object in getattr(from_item, f).all():
             # Create new related object
-            new_related_object, related_object = create_by_shallow_copy(request, related_object)
-            deep_copy(request, related_object, new_related_object)
+            new_related_object, related_object = create_by_shallow_copy(request, related_object, exclude, options)
+            deep_copy(request, related_object, new_related_object, exclude, options)
             getattr(to_item, f).add(new_related_object)
 
     m2mr_fields = [(x.name, x.field.name) for x in from_item._meta.get_fields() if type(x) == ManyToManyRel]
     for f, r in m2mr_fields:
         for related_object in getattr(from_item, f).all():
             # Create new related object
-            new_related_object, related_object = create_by_shallow_copy(request, related_object)
-            deep_copy(request, related_object, new_related_object)
+            new_related_object, related_object = create_by_shallow_copy(request, related_object, exclude, options)
+            deep_copy(request, related_object, new_related_object, exclude, options)
             getattr(to_item, f).add(new_related_object)
 
     # # Upstream manytomany fields still copied as links - have to link to new items
@@ -728,7 +728,7 @@ def deep_copy(request, from_item, to_item, exclude=[]):
     return from_item, to_item
 
 
-def copy_item(request, from_item, options, exclude=[]):
+def copy_item(request, from_item, exclude=[], options=[]):
     item_type = type(from_item).__name__.lower()
     try:
         item_model = ContentType.objects.get(app_label='main', model=item_type)
@@ -737,45 +737,19 @@ def copy_item(request, from_item, options, exclude=[]):
         messages.error(request, "{t}: {a}".format(t=type(e).__name__, a=e.args))
         return redirect('main:error')
 
-    from_item, to_item = create_by_shallow_copy(request, from_item, exclude)
+    from_item, to_item = create_by_shallow_copy(request, from_item, exclude, options)
     # Resetting the "from_item" after the copy as the pointers are changed by the change in pk?
 
     if options == 'link':
-        from_item, to_item = link_copy(request, from_item, to_item, exclude)
+        from_item, to_item = link_copy(request, from_item, to_item, exclude, options)
 
     if options == 'deep':
-        from_item, to_item = deep_copy(request, from_item, to_item, exclude)
+        from_item, to_item = deep_copy(request, from_item, to_item, exclude, options)
 
     return from_item, to_item
 
 
-def delete_item(request, item, options):
-    item_type = type(item).__name__.lower()
-    try:
-        item_model = ContentType.objects.get(app_label='main', model=item_type)
-    except Exception as e:
-        messages.error(request, "Could not find class with name '{t}'".format(t=item_type))
-        messages.error(request, "{t}: {a}".format(t=type(e).__name__, a=e.args))
-        return redirect('main:error')
-
-    if options == 'deep':
-        m = delete_deep(request, item)
-    else:
-        m = item.delete()
-
-    return m
-
-
-def delete_deep(request, item):
-    """
-    Deletes this item as well as any downstream items included in it.
-    :param request: 
-    :param item: 
-    :return: 
-    """
-
-    if item.owner != request.user.person:
-        return "{} skipped - not yours to delete".format(item.name)
+def detach_links(request, item):
 
     m2o_fields = [x.name for x in item._meta.get_fields() if type(x) == ManyToOneRel]
     if 'used_by' in m2o_fields:
@@ -783,13 +757,52 @@ def delete_deep(request, item):
 
     for f in m2o_fields:
         for related_object in getattr(item, f).all():
-            # Create new related object
+            # Remove related objects
+            detach_links(request, related_object)
+
+    m2mr_fields = [(x.name, x.field.name) for x in item._meta.get_fields() if type(x) == ManyToManyRel]
+    for f, r in m2mr_fields:
+        for related_object in getattr(item, f).all():
+            # Remove related object
+            detach_links(request, related_object)
+
+    item.owner = Person.objects.get(user__username="trash")
+    item.save()
+
+    return "Could not delete completely as the item is linked.  It has been moved to the recycle bin."
+
+
+def delete_item(request, item, options):
+
+    if options == 'deep':
+        m = delete_deep(request, item)
+    else:
+        m = item.delete()
+    return m
+
+
+def delete_deep(request, item):
+
+    if item.owner != request.user.person:
+        return "{} skipped - not yours to delete".format(item.name)  # not actually displayed anywhere ... ?
+
+    if item.used_by.count() > 0:
+        m = detach_links(request, item)
+        return m
+
+    m2o_fields = [x.name for x in item._meta.get_fields() if type(x) == ManyToOneRel]
+    if 'used_by' in m2o_fields:
+        m2o_fields.remove('used_by')
+
+    for f in m2o_fields:
+        for related_object in getattr(item, f).all():
+            # Remove related objects
             delete_deep(request, related_object)
 
     m2mr_fields = [(x.name, x.field.name) for x in item._meta.get_fields() if type(x) == ManyToManyRel]
     for f, r in m2mr_fields:
         for related_object in getattr(item, f).all():
-            # Create new related object
+            # Remove related object
             delete_deep(request, related_object)
 
     return item.delete()
